requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
bcrypt==5.0.0
bech32==1.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cosmpy==0.11.2
distlib==0.4.0
ecdsa==0.19.1
filelock==3.20.0
frozenlist==1.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
idna==3.11
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
multidict==6.7.0
platformdirs==4.5.0
propcache==0.4.1
protobuf==5.29.5
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
PyNaCl==1.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
six==1.17.0
sortedcontainers==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uagents==0.22.10
uagents-core==0.3.11
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.35.4
yarl==1.22.0
